ralph: write ranged rows as RANGES, and clear the OTTO wasm warnings (#152)
Two things, both surfaced by giving the wasm builds warning flags.

nesm.mps wrote a right-hand side of 25.699996999999996, which needs 18
characters to read back as the same double. Field 4 of a fixed-format
record is 12 columns, and no shorter decimal maps to that double -- I
checked rather than assumed -- so no change to how the value was printed
could make the file valid. The cause had to go instead.

The value was never in nesm. mps_reader.c expands a ranged row into two
constraints, Ax >= b - |r| and Ax <= b, because the model carries no
range concept. nesm's AP1P21 has b = 58.799988 and r = 33.099991, and
that subtraction is exactly 25.699996999999996. Writing the expansion
back out is correct but lossy: nesm came out as 751 rows instead of 663,
and the computed difference was unrepresentable.

The writer now folds the pair back into one row plus a RANGES entry. Two
properties keep that honest:

  - It merges only when the result round-trips exactly. The pair is
    recognised by adjacency, opposite senses and an identical coefficient
    vector, and merged only if hi - |hi - lo| reproduces lo bit-for-bit.
    Floating point does not promise that, and a merge that moved a
    constraint by an ulp would be a worse defect than a long line. Where
    it does not hold, both rows are written as before.

  - The range value is searched rather than computed. Only hi - |r| == lo
    has to hold and many decimals satisfy it; the raw difference can
    itself need 18 characters, where the shortest qualifying decimal is
    the source's own. nesm has ranges that compute to 50.800003000000004
    and are written as 50.800003.

nesm now writes 663 rows, matching its source, and the whole NETLIB set
round-trips through glpsol:

  fixed-format accepted:  84 / 84   (was 83)
  objective identical:    84 / 84

test_write_mps_ranges pins it and fails against the previous writer.

The four OTTO-code warnings are also fixed. tableau_invalidate_rc and
smoothstepf were dead static inlines and are deleted. The two
sign-compares were both integer-overflow guards comparing an int against
a size_t, where a negative value would have wrapped to something enormous
and tripped the guard by accident rather than by intent -- lap.c now
rejects a negative nnz explicitly, and velo_wasm.c casts. No OTTO-code
unused-function or sign-compare warnings remain in the wasm builds; the
247 that do are miniz's own, compiled in its own translation units.

+++ velo/wasm/src/velo_wasm.c
@@ -129,7 +129,10 @@ WasmRoute* wasm_route(VLGraph* graph,
 
         // Copy coordinates (with overflow check)
         if (route.num_coords > 0 && route.coords) {
-            if (route.num_coords <= SIZE_MAX / (2 * sizeof(double))) {
+            /* num_coords is int and already known positive above; the
+             * cast makes the comparison signed-to-signed explicit rather than
+             * relying on the implicit conversion. */
+            if ((size_t)route.num_coords <= SIZE_MAX / (2 * sizeof(double))) {
                 result->coords = malloc((size_t)route.num_coords * 2 * sizeof(double));
                 if (result->coords) {
                     for (int i = 0; i < route.num_coords; i++) {
